Testing the rover
Is to be hooked up to an h-bridge though a quad nand gate, so that only one output can control the direction of the motors on the other side of the h-bridge. Input 9 is used to control the direction of the motors, while input 10 turns the motors on and off. Variables were created for the outputs to the h-bridge for easy changing in case the rover version we have has different outputs hooked up to the h-bridge than the schematic I looked at, which can be found @ "http://www.robotshop.com/media/files/pdf/dfrobotshop-rover-schematic.pdf".