Arduino port of Python code from autonomous car #1
This file is the first try at porting over the logic from our Python file controlling the autonomous car. This is the first draft, and includes code for two sensors, with code for a third commented out for the most part. Also, made to utilize the analog output of the higher quality ultrasonic sensors we have access to, but can easily be adapted to fit the cheapo ones as well.